Login - split out StatsService and architected like User, Measurements
* With the services supplying data for the components that through the
magic of binding auto-update the view

TODO: Make the different stats in header and sidebar show different stats
* Header is ALL data (in current view?)
* Sidebar is just 'selected' data, which I think will just be from
selecting a user
