+++ src/app/stats/stats.header.service.ts
@@ -1,0 +1,9 @@
+import {StatsServiceBase} from './stats.base.service';
+import {Injectable} from '@angular/core';
+
+/**
+ * The purpose of this class is to create an injectable Stats service to use in header.
+ */
+@Injectable()
+export class HeaderStatsService extends StatsServiceBase {
+}
