+++ src/app/stats/stats.sidebar.service.ts
@@ -1,0 +1,10 @@
+import {StatsServiceBase} from './stats.base.service';
+import {Injectable} from '@angular/core';
+
+/**
+ * The purpose of this class is to create an injectable Stats service to use in sidebar.
+ */
+@Injectable()
+export class SidebarStatsService extends StatsServiceBase {
+
+}
